Themal Energy API
Got a good start on a system that is heavily inspired by the CoFH
Redstone Flux system.

This did require a few changes to machine, but nothing major.

Still needs a lot of work and testing.

+++ src/main/java/redgear/core/tile/TileEntityElectricMachine.java
@@ -24,7 +24,7 @@ public abstract class TileEntityElectricMachine extends TileEntityMachine implem
 	 * @return true if there is enough power, false if there is not
 	 */
 	@Override
-	protected final boolean tryUseEnergy(long energyUse) {
+	protected final boolean tryUseEnergy(int energyUse) {
 		if (energy > energyUse) {
 			energy -= energyUse;
 			return true;
